rename: menuNames.ts→MenuItems.ts
